requirements.txt
flask
flask-cors
pdfplumber
python-docx
python-dotenv
google-generativeai
